% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseline.R
\name{available}
\alias{available}
\title{Create a available matrix to use in the site select function.}
\usage{
available(data, cells)
}
\arguments{
\item{data}{raw data matrix of combined raster layers.}

\item{cells}{refers to the cells already chosen.}
}
\value{
available matrix of the original raster stack data.
}
\description{
Create a available matrix to use in the site select function.
}
