% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseline.R
\name{baseline}
\alias{baseline}
\title{Create a baseline matrix to use in the site select function.}
\usage{
baseline(data, cells)
}
\description{
Produces a matrix of the original cell values.
}
\details{
df' refers to the raw data matrix of combined raster layers.

'cells' refers to the cells already chosen.
}
