% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cell_combine.R
\name{cell_combine}
\alias{cell_combine}
\title{Combine individual cells into a list to use in the site_select() function.}
\usage{
cell_combine(...)
}
\description{
input any number of cell objects to combine into a list.
}
