% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cell_wrap.R
\name{cell_wrap}
\alias{cell_wrap}
\title{This function wraps two numeric values into a spatial object and returns a cell value corresponding to sampling extent}
\usage{
cell_wrap(x, y, raster_stack)
}
\arguments{
\item{x}{latitude value}

\item{y}{longitude value}

\item{raster_stack}{raster stack.}
}
\value{
a cell value corresponding to a raster stack from a given set of coordinates.
}
\description{
This function wraps two numeric values into a spatial object and returns a cell value corresponding to sampling extent
}
