% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_prop.R, R/log_to_prop.R
\name{class_prop}
\alias{class_prop}
\title{This function takes a classified logical raster, like land cover data (e.g., NLCD), masks a desired class, and converts it to a continuous data.}
\usage{
class_prop(rast, cls, extent, shell)

class_prop(rast, cls, extent, shell)
}
\arguments{
\item{rast}{the original classified logical raster.}

\item{cls}{the class values for the user's desired variable.
For instance, a single value (e.g, 21) could represent low urban cover in a classified logical raster.
Or, enter c(21, 22, 23) for 'cls' to mask low, medium, and high urban cover under a single raster object.}

\item{extent}{the 'extent' parameter crops the logical raster to a desired spatial constraint, either from the sample_ext() function or another vector boundary.}

\item{shell}{raster template for the 'shell' parameter that includes a desired number of columns and rows, extent, and crs.

#' Example - for raster 'x' :
shell = rast(ncol=ncol(x), nrow=nrow(x), ext = ext(x), crs = crs(x))
Note: To easily create a raster template copy information from an existent raster.}
}
\value{
the proportion of the occurrence of the raster is calculated in each cell of the raster template and returns a continuous raster.

the proportion of the occurrence of the raster is calculated in each cell of the raster template and returns a continuous raster.
}
\description{
This function takes a classified logical raster, like land cover data (e.g., NLCD), masks a desired class, and converts it to a continuous data.

This function takes a classified logical raster, like land cover data (e.g., NLCD), masks a desired class, and converts it to a continuous data.
}
