% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clim_load.R
\name{clim_load}
\alias{clim_load}
\title{this function takes NetCDF data which contains monthly means condensed into a single file.
clim_load() takes an nc file and uses climetric to calculate annual means for the geospatial analysis in other function of SiteSelectR.
the extent parameter refers to the sampling extent - see the sample_ext() function.
the extent crops new raster of annual averages to the desired sampling extent.}
\usage{
clim_load(nc, extent)
}
\arguments{
\item{nc}{NetCDF file}

\item{extent}{sampling extent to crop the raster.}
}
\value{
a single cropped object of mean annual climate that can be saved as a TIFF file for further geospatial analysis.
}
\description{
only works with the climetric package:
devtools::install_github("elizagrames/climetric")
}
