% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comparison.R
\name{comparison}
\alias{comparison}
\title{Create a comparison  matrix to use in the site select function.}
\usage{
comparison(data, cells)
}
\description{
Produces a matrix of the original cell values except those chosen already for sampling sites.
}
\details{
'df' refers to the raw data matrix of combined raster layers.

'cells' refers to the cells already chosen.
}
