% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_sites.R
\name{map_sites}
\alias{map_sites}
\title{Uses the cells produced from site_select and plots them on the first layer of
the raster stack.}
\usage{
map_sites(cells, stack)
}
\arguments{
\item{cells}{list of cells produced from site_select.}

\item{stack}{raster stack data.}
}
\value{
a plot of cells chosen from site_select.
}
\description{
Uses the cells produced from site_select and plots them on the first layer of
the raster stack.
}
