% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/max_cell.R
\name{max_cell}
\alias{max_cell}
\title{This function chooses the maximum different site to the starting location based on the MD.}
\usage{
max_cell(distance_matrix)
}
\arguments{
\item{distance_matrix}{calculated from rast_distance() function.}
}
\value{
a cell in the distance matrix with maximum dissimilarity (i.e., highest value).
}
\description{
This function chooses the maximum different site to the starting location based on the MD.
}
