% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_distance_fil.R
\name{new_distance_fil}
\alias{new_distance_fil}
\title{This function creates a new matrix using the available (with selected cells) and selected (without selected cells) matrices.
Modified for the filter function - site_select_fil - where the resulting dm is masked by the filter layer.}
\usage{
new_distance_fil(data, cells, filter_val)
}
\arguments{
\item{data}{the raw data matrix of combined raster layers.}

\item{cells}{refers to the cells already selected, or legacy sites that might exist and converted to cells that correspond to the raster stack.}
}
\value{
a distance matrix that measures the dissimilarity between the original distance matrix - see rast_distance() function - and the original distance matrix minus already selected cells.
}
\description{
This function creates a new matrix using the available (with selected cells) and selected (without selected cells) matrices.
Modified for the filter function - site_select_fil - where the resulting dm is masked by the filter layer.
}
