% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_sites.R
\name{plot_sites}
\alias{plot_sites}
\title{This function uses the cells chosen in the site_select() function.
Converts cells to geographic coordinates and plots on the first layer of the raster stack.}
\usage{
plot_sites(cells, stack)
}
\description{
'cells' refers to the output of site_select().
}
\details{
'stack', refers to the raster stack used as a reference to extract the geographic locations.
}
