% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rast_collect.R
\name{rast_collect}
\alias{rast_collect}
\title{This function collects a set of rasters into a list.}
\usage{
rast_collect(...)
}
\arguments{
\item{...}{any number of raster images.}
}
\value{
a list of rasters.
}
\description{
This function collects a set of rasters into a list.
}
