% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/first_distance.R, R/rast_distance.R
\name{rast_distance}
\alias{rast_distance}
\title{The rast_distance() function measures the initial dissimilarity present in the raster stack data without considering any chosen sampling sites.}
\usage{
rast_distance(data)

rast_distance(data)
}
\description{
This function uses the Mahalanobis distance calculation to measure multivariate dissimilarity.

This function uses the Mahalanobis distance calculation to measure multivariate dissimilarity.
}
\details{
'data' refers to the product of the rast_matrix() function - the matrix of numeric values extracted from the stack of raster images.

The covariance matrix included in the distance calculation uses an added diagonal matrix of 0.01 to avoid singularity issues.

The product from the first_distance() function should be applied to the __ funciton.

'data' refers to the product of the rast_matrix() function - the matrix of numeric values extracted from the stack of raster images.

The covariance matrix included in the distance calculation uses an added diagonal matrix of 0.01 to avoid singularity issues.

The product from the first_distance() function should be applied to the __ funciton.
}
