% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rast_distance_wt.R
\name{rast_distance_wt}
\alias{rast_distance_wt}
\title{modified version of rast_distance() and the wmahalanobis function from the WMDB package (Wu, 2012).
'weighted_layer' refers to the layer(s) in the raster stack to be weighted.
use c() to combine multiple layers before using them in the 'weighted layer' parameter.}
\usage{
rast_distance_wt(stack, weighted_layer, magnitude)
}
\description{
'magnitude' should be a weight value assigned to focal layers - relative to all other layers which are by default weighted at 0.001.
A large weight value would be >0.5.
}
