% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rast_stack.R
\name{rast_stack}
\alias{rast_stack}
\title{This function converts the resampled raster list into a geospatial object.}
\usage{
rast_stack(resampled, extent)
}
\description{
Input the product of the raster_match() function.
}
