% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_ext.R
\name{sample_ext}
\alias{sample_ext}
\title{Create a sampling circle around a central location}
\usage{
sample_ext(x, y, radius, crs)
}
\arguments{
\item{x}{Latitude of the center point.}

\item{y}{Longitude of the center point.}

\item{radius}{Desired radius of the sampling circle (in the units of the CRS).}

\item{crs}{EPSG code defining the coordinate reference system (CRS)}
}
\value{
A polygon (circle) representing the sampling extent. Used in the site_select() function.

Example:
sample_ext(lat, long, 100000 #meters, crs=4326)
}
\description{
This function is meant for users who want to calculate a sampling circle
around a central point. Whether using \code{sample_ext} directly or supplying a pre‑loaded shapefile, a sample extent is required for \code{site_select}.
}
