% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/site_select.R
\name{site_select}
\alias{site_select}
\title{This function creates a network of field sites based on multivariate GIS data and a starting
location.}
\usage{
site_select(long, lat, num, stack)
}
\arguments{
\item{long}{longitude value for the first site.}

\item{lat}{latitude value for the first site.}

\item{num}{number of sites to create the network.}

\item{stack}{raster stack of raster layers. Product of the rast_stack function.}
}
\value{
list of selected cells.
}
\description{
This function creates a network of field sites based on multivariate GIS data and a starting
location.
}
