% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/site_select_fil.R
\name{site_select_fil}
\alias{site_select_fil}
\title{This function is a modified version of site_select that is capable of masking to a set of vectorized polygons within the sampling extent (e.g., public land).
Masking occurs either before or after the Mahalanobis calculation.}
\usage{
site_select_fil(long, lat, num, stack, filter, type)
}
\arguments{
\item{long}{longitude value for the first site.}

\item{lat}{latitude value for the first site.}

\item{num}{number of sites to create the network.}

\item{stack}{raster stack of raster layers. Product of the rast_stack function.}

\item{type}{mask data either 'after' or 'before' the initial Mahalanobis distance calculation.}
}
\value{
list of selected cells.
}
\description{
This function is a modified version of site_select that is capable of masking to a set of vectorized polygons within the sampling extent (e.g., public land).
Masking occurs either before or after the Mahalanobis calculation.
}
