% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vect_to_rast.R
\name{vect_to_rast}
\alias{vect_to_rast}
\title{This function takes vectorized spatial data (e.g., points and lines).}
\usage{
vect_to_rast(vect, extent, res)
}
\description{
The vector is cropped to a desired extent - either from the sample_ext() function or an existing polygon.
}
\details{
The vector is projected to
}
