% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xy_wrap.R
\name{xy_wrap}
\alias{xy_wrap}
\title{similar to cell_wrap, xy_wrap takes a cell and produces a set of coordinates that are usable for site_select()}
\usage{
xy_wrap(cell, raster_stack)
}
\arguments{
\item{cell}{a list of cells.}

\item{raster_stack}{raster stack that corresponds to the geographic coordinates of the cells.}
}
\value{
list of geographic coordinates corresponding to the cells in the raster stack.
}
\description{
similar to cell_wrap, xy_wrap takes a cell and produces a set of coordinates that are usable for site_select()
}
