% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseline.R
\name{baseline}
\alias{baseline}
\title{Create a baseline matrix to use in the site select function.}
\usage{
baseline(data, cells)
}
\arguments{
\item{data}{raw data matrix of combined raster layers.}

\item{cells}{refers to the cells already chosen.}
}
\value{
baseline matrix of the original raster stack data.
}
\description{
Create a baseline matrix to use in the site select function.
}
