% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cell_wrap.R, R/xy_wrap.R
\name{cell_wrap}
\alias{cell_wrap}
\title{This function wraps two numeric values into a spatial object and returns a cell value corresponding to sampling extent}
\usage{
cell_wrap(cell, raster_stack)

cell_wrap(cell, raster_stack)
}
\description{
This function wraps two numeric values into a spatial object and returns a cell value corresponding to sampling extent

similar to cell_wrap, xy_wrap takes a cell and produces a set of coordinates that are usable for site_select()
}
