% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comparison.R
\name{comparison}
\alias{comparison}
\title{Create a comparison  matrix to use in the site select function.}
\usage{
comparison(data, cells)
}
\arguments{
\item{data}{data matrix of combined raster layers.}

\item{cells}{refers to the cells already chosen.}
}
\value{
a comparison matrix for the site_select() function. A matrix of data param with selected cells excluded.
}
\description{
Produces a matrix of the original cell values except those chosen already for sampling sites.
}
