% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/max_cell.R
\name{max_cell}
\alias{max_cell}
\title{Choose the maximum different site to the starting location.}
\usage{
max_cell(distance_matrix)
}
\description{
Using an x, y coordinate - chooses the cell from the original distance matrix with the highest dissimilarity to the starting location.
}
\details{
'matrix' refers to the output from the rast_mat() function - the raster stack resampled and in matrix-value form.

See: first_distance() function for original distance matrix.
}
