% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_distance.R
\name{new_distance}
\alias{new_distance}
\title{Create a new matrix using the baseline (with chosen cells) and comparison (without chosen cells) matrices.}
\usage{
new_distance(data, cells)
}
\description{
calculates Mahalanobis distance between a comparison and baseline set of cells.
}
\details{
'data' refers to the raw data matrix of combined raster layers.

'cells' refers to the cells already chosen
}
