% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_distance.R
\name{new_distance}
\alias{new_distance}
\title{This function creates a new matrix using the available (with selected cells) and selected (without selected cells) matrices.}
\usage{
new_distance(data, cells)
}
\arguments{
\item{data}{the raw data matrix of combined raster layers.}

\item{cells}{refers to the cells already selected, or legacy sites that might exist and converted to cells that correspond to the raster stack.}
}
\value{
a distance matrix that measures the dissimilarity between the original distance matrix - see rast_distance() function - and the original distance matrix minus already selected cells.
}
\description{
This function creates a new matrix using the available (with selected cells) and selected (without selected cells) matrices.
}
