% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_distance_wt.R
\name{new_distance_wt}
\alias{new_distance_wt}
\title{modified version of new_distance function, based on the wmahalanobis function in the WMDB package (Wu, 2012).
'cells' refers to the already selected cells, 'dat' to the raster matrix.}
\usage{
new_distance_wt(cells, dat)
}
\description{
modified version of new_distance function, based on the wmahalanobis function in the WMDB package (Wu, 2012).
'cells' refers to the already selected cells, 'dat' to the raster matrix.
}
