% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rast_collect.R
\name{rast_collect}
\alias{rast_collect}
\title{Used to collect a set of rasters into a list to be used in the smooth_raster() function.}
\usage{
rast_collect(...)
}
\description{
Input any variable number of rasters by their assigned object name.
}
