% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rast_match.R
\name{rast_match}
\alias{rast_match}
\title{Resample raster data to the same extent and resolution.}
\usage{
rast_match(rast_list, reference_layer, extent)
}
\arguments{
\item{rast_list}{List of rasters from the rast_collect function.}

\item{reference_layer}{A target layer for resampling. Written as a subset of
the object.}

\item{extent}{The product of sample_ext - a vectorized polygon to crop the rasters.}
}
\description{
Resample raster data to the same extent and resolution.
}
