% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rast_match.R
\name{rast_match}
\alias{rast_match}
\title{Align raster data to the same extent and resolution.}
\usage{
rast_match(raster_list, reference_layer, extent)
}
\description{
'reference_layer' refers to the raster that others will be projected to.
'extent' refers to the spatial boundaries that the rasters will be cropped to.
}
\details{
Rasters need to be aligned and projected to the same resolution and CRS before condensing into a raster stack and before calculating a distance matrix.
}
