% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_ext.R
\name{sample_ext}
\alias{sample_ext}
\title{The functions is meant for those interested in creating a simple sampling circumfrence around a central location as sampling extent.}
\usage{
sample_ext(x, y, radius)
}
\description{
The sampling extent is used in the raster_match() function.
}
\details{
x=latitude, y=longitude, 'radius' refers to the radius of the sampling circumfrence.
}
