% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/site_select.R
\name{site_select}
\alias{site_select}
\title{Choose a set of sites from the new distance matrix.
see comparison(), baseline(), & new_distance() functions.}
\usage{
site_select(long, lat, num, stack)
}
\description{
'num' refers the number of rounds of site selection added to existing cells/sites.
}
\details{
'cells' refers to the starting location and cell with the highest dissimilarity

'stack' refers to the output from the rast_stack() function - the raster stack resampled to the same extent and resolution.
}
