% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/site_select_wt.R
\name{site_select_wt}
\alias{site_select_wt}
\title{This function chooses a set of sites from the new distance matrix.}
\usage{
site_select_wt(long, lat, num, stack, weighted_layer, magnitude)
}
\arguments{
\item{long}{the starting location.}

\item{lat}{the starting location.}

\item{num}{the number of rounds of site selection added to existing cells/sites.}

\item{stack}{refers to the output from the rast_stack() function - the raster stack resampled to the same extent and resolution.}

\item{weighted_layer}{is the layer number in the raster stack to be weighted. This can be multiple numbers in c().}

\item{magnitude}{if magnitude = NULL, default is 0.1 weight.}
}
\description{
This function chooses a set of sites from the new distance matrix.
}
