% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/site_val.R
\name{site_val}
\alias{site_val}
\title{This function creates a matrix with the values of the sites at each layer of raster data.}
\usage{
site_val(sites, stack)
}
\arguments{
\item{sites}{the site coordinates from the plot_sites() function.}

\item{stack}{raster stack.}
}
\value{
the values of each variable in the raster stack at each selected cell during site selection.
}
\description{
This function creates a matrix with the values of the sites at each layer of raster data.
}
