% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis_cor.R
\name{vis_cor}
\alias{vis_cor}
\title{This function makes a simple correlation plot to visualize the collinearity between env. variables at selected sites.}
\usage{
vis_cor(values)
}
\arguments{
\item{values}{the site values produced from the site_val() function.}
}
\value{
a matrix of correlation values for the raster data and a heat map visualizing environmental collinearity between site values.
}
\description{
This function makes a simple correlation plot to visualize the collinearity between env. variables at selected sites.
}
