% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis_rep.R
\name{vis_rep}
\alias{vis_rep}
\title{This function takes two layers of raster data to compare to the site values.}
\usage{
vis_rep(values, stack, layer1, layer2)
}
\description{
A good indicator of representative site selection is broad coverage of the mean, min, and max raster data.
}
\details{
Input the raster stack - output of rast_stck() - and site values - output of site_val().

'layer1' and 'layer2' refers to the y and x axis of the plot and the subset of layers used in the raster stack.
}
