% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wt_distance.R
\name{wt_distance}
\alias{wt_distance}
\title{modified from the wmahalanobis function from the WMDB package}
\usage{
wt_distance(data, center, cov, weight)
}
\description{
Wu, B., & Csárdi, G. (2010). WMDB: Discriminant Analysis Methods by Weight Mahalanobis Distance and Bayes* (Version 1.0) R package. CRAN. https://CRAN.R-project.org/package=WMDB
}
